refactor(uilanding): merge latency handlers into impairments.py
